# Conflicts:
#   frontend/src/setupProxy.js

+++ backend/ssapilogue/src/main/java/com/ssafy/ssapilogue/api/controller/UserController.java
@@ -1,15 +1,21 @@
 package com.ssafy.ssapilogue.api.controller;
 
+import com.ssafy.ssapilogue.api.config.JwtProperties;
 import com.ssafy.ssapilogue.api.dto.request.LoginUserReqDto;
 import com.ssafy.ssapilogue.api.dto.request.SignupUserReqDto;
 import com.ssafy.ssapilogue.api.dto.request.UpdateUserReqDto;
 import com.ssafy.ssapilogue.api.dto.response.FindUserResDto;
+import com.ssafy.ssapilogue.api.dto.response.LoginUserResDto;
 import com.ssafy.ssapilogue.api.dto.response.SignupUserResDto;
+import com.ssafy.ssapilogue.api.exception.CustomException;
+import com.ssafy.ssapilogue.api.exception.ErrorCode;
 import com.ssafy.ssapilogue.api.service.JwtTokenProvider;
 import com.ssafy.ssapilogue.api.service.UserService;
 import com.ssafy.ssapilogue.api.util.MediaUtils;
 import com.ssafy.ssapilogue.core.domain.User;
 import com.ssafy.ssapilogue.core.repository.UserInfoRepository;
+import com.ssafy.ssapilogue.core.repository.UserRepository;
+import io.jsonwebtoken.ExpiredJwtException;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
 import io.swagger.annotations.ApiParam;
@@ -24,9 +30,8 @@ import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
 
 import javax.servlet.http.HttpServletRequest;
-import java.io.FileInputStream;
-import java.io.InputStream;
-import java.nio.charset.StandardCharsets;
+import javax.servlet.http.HttpServletResponse;
+import java.io.IOException;
 import java.util.HashMap;
 import java.util.Map;
 
@@ -40,6 +45,8 @@ public class UserController {
 
     private final UserInfoRepository userInfoRepository;
 
+    private final UserRepository userRepository;
+
     private final JwtTokenProvider jwtTokenProvider;
 
     @Value("${profileImg.path}")
@@ -51,68 +58,70 @@ public class UserController {
             @RequestBody @ApiParam(value = "회원가입 유저 정보", required = true) SignupUserReqDto signupUserReqDto) throws Exception {
         Map<String, Object> result = new HashMap<>();
         HttpStatus httpStatus = null;
-        try {
-            SignupUserResDto signupUserResDto = userService.signup(signupUserReqDto);
-            result.put("userinfo", signupUserResDto);
-            httpStatus = HttpStatus.OK;
-            result.put("message", "success");
-            System.out.println("success");
-            String token = jwtTokenProvider.createToken(signupUserResDto.getEmail(), signupUserResDto.getUserIdentity());
-            result.put("token", token);
-        }
-        catch(Exception e) {
-            e.printStackTrace();
-            httpStatus = HttpStatus.INTERNAL_SERVER_ERROR;
-            result.put("message", "error");
-            System.out.println("error");
-        }
+
+        User user = userService.signup(signupUserReqDto);
+
+        httpStatus = HttpStatus.OK;
+        result.put("message", "success");
+
+        SignupUserResDto signupUserResDto = new SignupUserResDto(user);
+        result.put("userinfo", signupUserResDto);
+
+        String token = jwtTokenProvider.createToken(signupUserResDto.getEmail(), signupUserResDto.getUserIdentity());
+        result.put("token", token);
+
+        String refreshToken = jwtTokenProvider.createRefreshToken(signupUserReqDto.getEmail());
+        result.put("refresh-token", refreshToken);
+        user.changeRefreshToken(refreshToken);
+        userRepository.save(user);
+
         return new ResponseEntity<Map<String, Object>>(result, httpStatus);
     }
 
     @PostMapping("/login")
     @ApiOperation(value = "로그인", notes = "로그인을 한다.")
     public ResponseEntity<Map<String, Object>> login(
-            @RequestBody @ApiParam(value = "로그인 유저 정보", required = true) LoginUserReqDto loginUserReqDto) {
+            @RequestBody @ApiParam(value = "로그인 유저 정보", required = true) LoginUserReqDto loginUserReqDto) throws Exception {
         Map<String, Object> result = new HashMap<>();
         HttpStatus httpStatus = null;
-        if (userInfoRepository.findByUserId(loginUserReqDto.getUserId()) == null) {
-            httpStatus = HttpStatus.NOT_ACCEPTABLE;
-            result.put("message", "ssapilogue를 이용할 수 없습니다.");
-            return new ResponseEntity<Map<String, Object>>(result, httpStatus);
-        }
-        try {
-            User loginUser = userService.login(loginUserReqDto);
-            httpStatus = HttpStatus.OK;
-            result.put("status", "SUCCESS");
-            String token = jwtTokenProvider.createToken(loginUser.getEmail(), loginUser.getUserIdentity());
-            result.put("token", token);
-        } catch (NullPointerException e) {
-            httpStatus = HttpStatus.OK;
-            result.put("status", "NO USER");
-        } catch (RuntimeException e) {
-            httpStatus = HttpStatus.INTERNAL_SERVER_ERROR;
-            result.put("status", "SEVER ERROR");
-        } catch (Exception e) {
-            e.printStackTrace();
-        }
+        User loginUser = userService.login(loginUserReqDto);
+
+        httpStatus = HttpStatus.OK;
+        result.put("status", "SUCCESS");
+
+        String token = jwtTokenProvider.createToken(loginUser.getEmail(), loginUser.getUserIdentity());
+        result.put("token", token);
+
+        String refreshToken = jwtTokenProvider.createRefreshToken(loginUserReqDto.getEmail());
+        result.put("refresh-token", refreshToken);
+        loginUser.changeRefreshToken(refreshToken);
+        userRepository.save(loginUser);
+
         return new ResponseEntity<Map<String, Object>>(result, httpStatus);
     }
 
+
+
     @PutMapping
     @ApiOperation(value = "회원정보 수정", notes = "회원정보를 수정한다.")
-    public ResponseEntity<Map<String, Object>> updateUser(
+    public ResponseEntity<Map<String, Object>> updateUser(HttpServletRequest request,
             @RequestBody @ApiParam(value = "회원정보 수정", required = true) UpdateUserReqDto updateUserReqDto) {
         Map<String, Object> result = new HashMap<>();
         HttpStatus httpStatus = null;
-        try {
-            userService.updateUser(updateUserReqDto);
-            httpStatus = HttpStatus.OK;
-            result.put("status", "SUCCESS");
-        } catch (Exception e) {
-            e.printStackTrace();
-            httpStatus = HttpStatus.INTERNAL_SERVER_ERROR;
-            result.put("status", "SERVER ERROR");
-        }
+
+        String token = jwtTokenProvider.resolveToken(request);
+        if (token == null) throw new CustomException(ErrorCode.NO_TOKEN);
+
+        String userEmail = jwtTokenProvider.getUserEmail(token);
+        if (userEmail == null) throw new CustomException(ErrorCode.WRONG_TOKEN);
+
+        User user = userRepository.findByEmail(userEmail);
+        if (user == null) throw new CustomException(ErrorCode.NO_USER);
+
+        userService.updateUser(user, updateUserReqDto);
+        httpStatus = HttpStatus.OK;
+        result.put("status", "SUCCESS");
+
         return new ResponseEntity<Map<String, Object>>(result, httpStatus);
     }
 
@@ -121,16 +130,17 @@ public class UserController {
     public ResponseEntity<Map<String, Object>> deleteUser(HttpServletRequest request) {
         Map<String, Object> result = new HashMap<>();
         HttpStatus httpStatus = null;
+
         String token = jwtTokenProvider.resolveToken(request);
+        if (token == null) throw new CustomException(ErrorCode.NO_TOKEN);
+
         String userEmail = jwtTokenProvider.getUserEmail(token);
-        try {
-            userService.deleteUser(userEmail);
-            httpStatus = HttpStatus.OK;
-            result.put("status", "SUCCESS");
-        } catch (Exception e) {
-            httpStatus = HttpStatus.INTERNAL_SERVER_ERROR;
-            result.put("status", "SERVER ERROR");
-        }
+        if (userEmail == null) throw new CustomException(ErrorCode.WRONG_TOKEN);
+
+        userService.deleteUser(userEmail);
+        httpStatus = HttpStatus.OK;
+        result.put("status", "SUCCESS");
+
         return new ResponseEntity<Map<String, Object>>(result, httpStatus);
     }
 
@@ -140,17 +150,16 @@ public class UserController {
         Map<String, Object> result = new HashMap<>();
         HttpStatus httpStatus = null;
         FindUserResDto findUser = null;
+
         String token = jwtTokenProvider.resolveToken(request);
+        if (token == null) throw new CustomException(ErrorCode.NO_TOKEN);
+
         String userEmail = jwtTokenProvider.getUserEmail(token);
-        try {
-            findUser = userService.findUserProfile(userEmail);
-            httpStatus = HttpStatus.OK;
-            result.put("status", "SUCCESS");
-        } catch (Exception e) {
-            e.printStackTrace();
-            httpStatus = HttpStatus.INTERNAL_SERVER_ERROR;
-            result.put("status", "SERVER_ERROR");
-        }
+        if (userEmail == null) throw new CustomException(ErrorCode.WRONG_TOKEN);
+
+        findUser = userService.findUserProfile(userEmail);
+        httpStatus = HttpStatus.OK;
+        result.put("status", "SUCCESS");
         result.put("user", findUser);
         return new ResponseEntity<Map<String, Object>>(result, httpStatus);
     }
@@ -163,17 +172,24 @@ public class UserController {
         HttpStatus httpStatus = null;
 
         String token = jwtTokenProvider.resolveToken(request);
-        String userEmail = jwtTokenProvider.getUserEmail(token);
+        if (token == null) throw new CustomException(ErrorCode.NO_TOKEN);
 
-        try {
-            String imageUrl = userService.updateImage(userEmail, file);
-            httpStatus = HttpStatus.OK;
-            result.put("imageUrl", imageUrl);
-        } catch (Exception e) {
-            e.printStackTrace();
-            httpStatus = HttpStatus.INTERNAL_SERVER_ERROR;
-            result.put("status", "SERVER ERROR");
-        }
+        String userEmail = jwtTokenProvider.getUserEmail(token);
+        if (userEmail == null) throw new CustomException(ErrorCode.WRONG_TOKEN);
+
+        String imageUrl = userService.updateImage(userEmail, file);
+        httpStatus = HttpStatus.OK;
+        result.put("imageUrl", imageUrl);
+        
         return new ResponseEntity<Map<String, Object>>(result, httpStatus);
+    }
+
+    @GetMapping("/refreshToken")
+    public void refreshToken(HttpServletRequest request, HttpServletResponse response) throws Exception {
+        String refreshToken = jwtTokenProvider.resolveRefreshToken(request);
+
+        if (refreshToken == null) throw new CustomException(ErrorCode.NO_REFRESH_TOKEN);
+        String newToken = userService.issueNewToken(refreshToken);
+        response.addHeader(JwtProperties.HEADER_STRING, newToken);
     }
 }
